fix: consolidate card fetching logic and remove redundant API endpoints
- Removed redundant `fetchFlashcards` API endpoint and related client-side functions.
- Consolidated card fetching logic within `CardContext` and `App` component to ensure data is loaded once.
- Updated `Content` and `DetailPage` components to rely on `CardContext` for card data.
- Modified `Ongoing` component to use directly compared answers instead of API validation.
- Cleaned up imports and ensured consistent data handling across components.
