+++ src/client/pages/App.tsx
@@ -6,15 +6,30 @@ import NewPage from '../components/layouts/GamePage/NewPage'
 import Ongoing from '../components/layouts/GamePage/Ongoing'
 import EndPage from '../components/layouts/GamePage/EndPage'
 import DetailPage from '../components/layouts/GamePage/DetailPage'
-import { useContext } from 'react'
+import { useContext, useEffect } from 'react'
 import { GameContext } from '../../api/GameContext'
 import { GameState } from '../../api/GameState'
+import { fetchCards } from '../../api/card'
+import { CardContext } from '../../api/CardContext'
 
 function App() {
-  const { state } = useContext(GameContext)
+  const { state: gameState } = useContext(GameContext)
+  const { dispatch: cardDispatch } = useContext(CardContext)
+
+  useEffect(() => {
+    const fetchData = async () => {
+      try {
+        const fetchedCards = await fetchCards()
+        cardDispatch({ type: 'SET_CARDS', payload: fetchedCards })
+      } catch (error) {
+        console.error(error)
+      }
+    }
+    fetchData()
+  }, [])
 
   const renderContent = () => {
-    switch (state.gameState) {
+    switch (gameState.gameState) {
       case GameState.NOT_STARTED:
         return <NewPage />
       case GameState.START:
