+++ src/server/controllers/cards.ts
@@ -40,6 +40,7 @@ export const deleteCard = (req: Request, res: Response) => {
   res.status(204).send()
 }
 
+// TODO backend but not over api
 export const fetchFlashcards = (_req: Request, res: Response) => {
   const cards = cardStore.getCards()
   const maxIndex = cards.length > 10 ? 10 : cards.length
