+++ src/server/routes/cards.ts
@@ -4,7 +4,6 @@ import * as cardsController from '../controllers/cards'
 const router = express.Router()
 
 router.get('/getCards', cardsController.getCards)
-router.get('/fetchFlashcards', cardsController.fetchFlashcards)
 router.get('/getCardById/:id', cardsController.getCardById)
 router.post('/addCard', cardsController.addCard)
 router.put('/updateCard/:id', cardsController.updateCard)
